Working on breaking out app from di core
