+++ module/di/bin/lib/util.ts
@@ -1,6 +1,6 @@
 import * as child_process from 'child_process';
 
-import { ApplicationConfig } from '../../src/types';
+import { ApplicationConfig } from '../../src/app/types';
 
 export interface CachedAppConfig extends ApplicationConfig {
   appRoot: string;
