+++ module/di/index.ts
@@ -2,3 +2,6 @@ export * from './src/decorator';
 export * from './src/error';
 export * from './src/registry';
 export * from './src/types';
+export * from './src/app/decorator';
+export * from './src/app/registry';
+export * from './src/app/types';
