+++ module/di/src/app/registry.ts
@@ -1,0 +1,51 @@
+import { Env, AppInfo } from '@travetto/base';
+import { ConfigSource } from '@travetto/config';
+
+import { ApplicationConfig } from './types';
+import { InjectionError } from '../error';
+import { DependencyRegistry } from '../registry';
+
+export class $ApplicationRegistry {
+  private applications = new Map<string, ApplicationConfig>();
+
+  register(app: string, config: ApplicationConfig) {
+    this.applications.set(app, config);
+  }
+
+  loadAllFromConfig() {
+    for (const entries of Object.values(ConfigSource.get('di.application') || {}) as string[][]) {
+      for (const entry of entries) {
+        require(entry);
+      }
+    }
+  }
+
+  getAll() {
+    return Array.from(this.applications.values());
+  }
+
+  async run(name: string, args: any[]) {
+    const config = this.applications.get(name);
+    if (!config) {
+      throw new InjectionError(`Application: ${name} does not exist`, 'notfound');
+    }
+    const inst = await DependencyRegistry.getInstance(config.target);
+    if (!Env.quietInit) {
+      console.log('Running application', name);
+      console.log('Configured', JSON.stringify({
+        app: AppInfo,
+        env: Env.toJSON(),
+        config: ConfigSource.get(''),
+      }, undefined, 2));
+    }
+    if (inst.run) {
+      await inst.run(...args);
+    }
+  }
+
+  onReset() {
+    this.applications.clear();
+  }
+}
+
+export const ApplicationRegistry = new $ApplicationRegistry();
