+++ module/di/src/decorator.ts
@@ -1,6 +1,6 @@
 import { Class } from '@travetto/registry';
 
-import { InjectableFactoryConfig, InjectableConfig, Dependency, ApplicationConfig, ApplicationParameter } from './types';
+import { InjectableFactoryConfig, InjectableConfig, Dependency } from './types';
 import { DependencyRegistry } from './registry';
 
 function extractSymbolOrConfig<T extends { qualifier?: Symbol }>(args: any[]) {
@@ -25,35 +25,6 @@ export function Injectable(...args: any[]): ClassDecorator {
 
     config.class = target;
     DependencyRegistry.registerClass(target, config as any as InjectableConfig<any>);
-    return target;
-  };
-}
-
-export type AppDecorator = Partial<ApplicationConfig> & {
-  paramMap?: {
-    [key: string]: Partial<ApplicationParameter> & { name?: never }
-  }
-  params?: never;
-};
-
-export function Application(
-  name: string,
-  config?: AppDecorator,
-  params?: (Partial<ApplicationParameter> & { name: string })[]
-): ClassDecorator {
-  return (target: Class | any) => {
-    const out: Partial<ApplicationConfig> = (config || {});
-    const paramMap = config && config.paramMap || {};
-
-    out.target = target;
-    out.name = name.replace(/(\s+|[^A-Za-z0-9\-_])/g, '-');
-    out.standalone = out.standalone === undefined || out.standalone; // Default to standalone
-
-    if (params) {
-      out.params = params.map(x => ({ ...x, ...(paramMap[x.name!] || {}), name: x.name! }) as ApplicationParameter);
-    }
-
-    DependencyRegistry.registerApplication(out.name, out as ApplicationConfig);
     return target;
   };
 }
